Movie Data Analysis
In this project, I examine a dataset obtained at https://www.kaggle.com/datasets/danielgrijalvas/movies.

This dataset contains data on four decades of movies. It required no real cleaning as it is a widely used dataset on kaggle.

That being said, using some basic commands in Python, Seaborn, pandas, and NumPy, I determine which factors have historically had the most correlation to gross revenue.

I began this project with the hypothesis that budget would have a high correlation with gross revenue. Interestingly, the number of user votes the movie received. 

Those looking to make their own movie could use this information to have a more profitable movie. That is to say, If one wants their movie to generate large sums of gross revenue, they should focus on obtaining a larger budget as well as directing the movie in such a way that it receives as many user votes if possible.

There are two potential issues with this:

First, "user votes" is left vague with no explanation as to where the votes are being cast, by whom, or for what purpose. Knowing these details would greatly enhance the usability of this data. 

Second, my analysis proves an undeniable correlation between budget and votes to gross revenue; however, they do not imply causation.